+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 **Backwards incompatible** due to required authentication when using `ProfileAvatar`
 
-* `PasswordResetEmail` now only throttled on `POST`.
+* `PasswordResetEmail` and `GetToken` now only throttled on `POST` requests.
 * Added `DELETE` method to `ProfileAvatar`.
 * `ProfileAvatar` now requires authentication.